requirements.txt
google-api-python-client
PyPDF2
transformers
flask
torch
python-dotenv
ollama